1requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
boto3==1.21.35
botocore==1.24.37
Django==4.0.4
django-appconf==1.0.5
django-ckeditor==6.3.0
django-compressor==3.1
django-cors-headers==3.11.0
django-countries==7.3.2
django-debug-toolbar==3.4.0
django-js-asset==2.0.0
django-jsoneditor==0.2.2
django-storages==1.12.3
gunicorn==20.1.0
jmespath==1.0.0
packaging==21.3
Pillow==9.1.0
psycopg2-binary==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.6
rcssmin==1.1.0
rjsmin==1.2.0
s3transfer==0.5.2
six==1.16.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
